pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "internship-assistant"
version = "0.1.0"
description = "CLI-first internship discovery and application assistance (Greenhouse + Lever), with AI enrichment and Google Sheets mirroring."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{name = "Community", email = "example@example.com"}]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "requests>=2.32.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.2.0",
  "python-dotenv>=1.0.1",
  "tenacity>=8.3.0",
  "diskcache>=5.6.3",
  "tqdm>=4.66.0",
  "google-auth>=2.33.0",
  "google-auth-oauthlib>=1.2.1",
  "google-api-python-client>=2.139.0",
  "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-requests>=2.32.0.20240712",
]
playwright = [
  "playwright>=1.46.0",
]

[project.scripts]
internship-assistant = "app.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["storage*", "tests*"]
